This is PURE kernel 3.0 reverted from history! Next step will be adding parts from kernel 5.5 till i find the bugs! i hope...

+++ arch/x86/mm/mmap.c
@@ -84,14 +84,14 @@ static unsigned long mmap_rnd(void)
 	/*
 	*  8 bits of randomness in 32bit mmaps, 20 address space bits
 	* 28 bits of randomness in 64bit mmaps, 40 address space bits
-       */
-       if (current->flags & PF_RANDOMIZE) {
-               if (mmap_is_ia32())
-                       rnd = get_random_int() % (1<<8);
-               else
-                       rnd = get_random_int() % (1<<28);
-       }
-       return rnd << PAGE_SHIFT;
+	*/
+	if (current->flags & PF_RANDOMIZE) {
+		if (mmap_is_ia32())
+			rnd = (long)get_random_int() % (1<<8);
+		else
+			rnd = (long)(get_random_int() % (1<<28));
+	}
+	return rnd << PAGE_SHIFT;
 }
 
 static unsigned long mmap_base(void)
